Bump version to 1.0.8
BFloat16 Compatibility Fix
- Fixed tensor type compatibility issues with audio processing nodes
- Input audio tensors are now converted from BFloat16 to Float32 for numpy compatibility
- Output audio tensors are explicitly converted to Float32 to ensure compatibility with downstream nodes
- Resolves "Got unsupported ScalarType BFloat16" errors when using voice cloning or saving audio

+++ nodes/base_vibevoice.py
@@ -478,8 +478,10 @@ class BaseVibeVoiceNode:
                     elif audio_tensor.dim() == 2:
                         audio_tensor = audio_tensor.unsqueeze(0)
                     
+                    # Convert to float32 for compatibility with downstream nodes (Save Audio, etc.)
+                    # Many audio processing nodes don't support BFloat16
                     return {
-                        "waveform": audio_tensor.cpu(),
+                        "waveform": audio_tensor.cpu().float(),
                         "sample_rate": 24000
                     }
